working on the tests
